backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
qdrant-client==1.7.1
openai==1.3.0
cohere==4.37
groq==0.4.0
pydantic==2.5.0
python-multipart==0.0.6
tiktoken==0.5.2

